Add multilingual locales and translations
