+++ i18n.js
@@ -1,0 +1,5 @@
+export default {
+  locales: ["en", "pt", "es", "ko", "zh-hans", "zh-hant", "ar"],
+  defaultLocale: "en",
+  messagesDir: "./messages"
+};
