+++ messages/ko.json
@@ -1,0 +1,212 @@
+{
+  "nav": {
+    "language": "언어 선택"
+  },
+  "a11y": {
+    "nav": "주요 내비게이션",
+    "whatsapp": "WhatsApp 채팅 열기"
+  },
+  "page": {
+    "self": {
+      "brand": "Wonnymed",
+      "nav": {
+        "solutions": "솔루션",
+        "about": "회사 소개",
+        "how": "진행 방식",
+        "compliance": "컴플라이언스",
+        "rfq": "RFQ",
+        "portal": "포털"
+      },
+      "hero": {
+        "titleA": "컴플라이언스와 ",
+        "titleB": "속도",
+        "titleC": "로 완성하는 임상 공급.",
+        "sub": "지혈제, 봉합사, 페이퍼유즈 드릴, 전문 더마 제품. 🇨🇳 중국(Med): 검증된 공급업체가 경쟁력 있는 가격과 관리된 품질을 제공합니다. 🇰🇷 한국(Beauty): 프리미엄 품질을 제공하는 선도 제조사. 24–48시간 이내 견적, 검증된 ANVISA/UDI 문서, 배송 SLA 관리.",
+        "ctaPrimary": "견적 요청",
+        "ctaSecondary": "컴플라이언스 보기",
+        "note": "공개 카탈로그가 없습니다. RFQ와 승인자를 위한 비공개 포털로 운영됩니다.",
+        "badgeTitle": "Verified 인증",
+        "badgeList": [
+          "ANVISA & UDI 등록",
+          "ISO 1345 검증",
+          "IFU/MSDS 첨부",
+          "로트 추적"
+        ],
+        "originBadges": [
+          {
+            "text": "China (Med)",
+            "flag": "🇨🇳",
+            "ariaLabel": "중국 국기",
+            "tone": "med"
+          },
+          {
+            "text": "Korea (Beauty)",
+            "flag": "🇰🇷",
+            "ariaLabel": "대한민국 국기",
+            "tone": "beauty"
+          }
+        ]
+      },
+      "metrics": [
+        { "k": "견적 SLA", "v": "24–48h" },
+        { "k": "OTIF", "v": ">= 95%" },
+        { "k": "루틴", "v": "비공개 포털" },
+        { "k": "컴플라이언스", "v": "ANVISA • UDI • ISO" },
+        { "k": "파트너 네트워크", "v": "🇨🇳 • 🇰🇷" }
+      ],
+      "solutionsTitle": "솔루션",
+      "lines": [
+        {
+          "title": "지혈제",
+          "desc": "완전한 문서와 함께하는 출혈 관리. 🇨🇳 중국(Med): 감사된 품질과 경쟁력 있는 가격.",
+          "icon": "🩺",
+          "badge": {
+            "text": "China (Med)",
+            "flag": "🇨🇳",
+            "ariaLabel": "중국 국기",
+            "tone": "med"
+          }
+        },
+        {
+          "title": "봉합사",
+          "desc": "흡수성/비흡수성 옵션과 호환성, IFU 제공. 🇨🇳 중국(Med): 감사된 품질과 경쟁력 있는 가격.",
+          "icon": "🧵",
+          "badge": {
+            "text": "China (Med)",
+            "flag": "🇨🇳",
+            "ariaLabel": "중국 국기",
+            "tone": "med"
+          }
+        },
+        {
+          "title": "드릴 PPU",
+          "desc": "페이퍼유즈, 중립 AFE, 가동률 ≥ 98%, 72시간 내 교체.",
+          "icon": "🛠️"
+        },
+        {
+          "title": "Derma Pro (Beauty)",
+          "desc": "프리미엄 임상 에스테틱 — 아래 시장에 중점 공급. 🇰🇷 한국(Beauty): 업계를 선도하는 제조사의 프리미엄 품질.",
+          "icon": "💠",
+          "badge": {
+            "text": "Korea (Beauty)",
+            "flag": "🇰🇷",
+            "ariaLabel": "대한민국 국기",
+            "tone": "beauty"
+          },
+          "rfqLineSlug": "dermato",
+          "focusMarkets": [
+            { "code": "br", "flag": "🇧🇷", "label": "브라질" },
+            { "code": "ae", "flag": "🇦🇪", "label": "UAE" },
+            { "code": "sa", "flag": "🇸🇦", "label": "사우디" },
+            { "code": "kw", "flag": "🇰🇼", "label": "쿠웨이트" }
+          ]
+        }
+      ],
+      "about": {
+        "eyebrow": "회사",
+        "title": "글로벌 거버넌스를 갖춘 임상 공급 헬스테크.",
+        "subtitle": "아시아 전역의 감사된 제조사와 외과·피부과 팀을 연결해 물류 예측 가능성, ANVISA/UDI 컴플라이언스, 엔터프라이즈급 구매 경험을 제공합니다.",
+        "missionTitle": "목표",
+        "mission": "임상공학 및 조달팀을 위해 전체 컴플라이언스, 투명한 문서, 모니터링되는 SLA로 중요한 공급을 보장합니다.",
+        "valuesTitle": "우리의 차별점",
+        "values": [
+          {
+            "title": "마찰 없는 컴플라이언스",
+            "desc": "견적 이전 ANVISA, UDI, ISO 1345, IFU/MSDS 실사와 유효기간 알림 제공."
+          },
+          {
+            "title": "감사된 네트워크",
+            "desc": "홍콩 허브와 브라질·걸프 지역 팀이 검사, 집하, 최종 배송을 조율합니다."
+          },
+          {
+            "title": "협업 포털",
+            "desc": "임상·구매·재무팀이 모든 로트를 전주기 추적과 함께 승인할 수 있는 워크플로우."
+          }
+        ],
+        "statsTitle": "거버넌스 & 커버리지",
+        "stats": [
+          { "value": "6+", "label": "2023/24년 완료된 공장 감사" },
+          { "value": "24–48h", "label": "표준 기술 비교 SLA" },
+          { "value": "≥95%", "label": "선제 모니터링이 뒷받침하는 OTIF 목표" }
+        ],
+        "quote": "\"엔드투엔드 추적 가능성으로 임상팀에 예측 가능성을 제공합니다.\"",
+        "quoteBy": "Wonnymed 리더십"
+      },
+      "askQuote": "견적 요청",
+      "howTitle": "진행 방식",
+      "howSteps": [
+        { "t": "요청", "d": "사양, 수량, 일정이 포함된 RFQ를 전송합니다." },
+        { "t": "검증", "d": "ANVISA/UDI/ISO, 호환성, 유효기간을 검토합니다." },
+        { "t": "납품", "d": "24–48시간 내 견적과 OTIF ≥ 95% 제공." }
+      ],
+      "complianceTitle": "컴플라이언스 & 문서",
+      "complianceDesc": "규제 투명성을 기본값으로 제공합니다. 모든 제안에 기술 자료와 추적이 포함됩니다.",
+      "complianceList": [
+        "ANVISA 및 UDI 등록",
+        "제조사 ISO 1345",
+        "최신 IFU/MSDS",
+        "로트/유효기간 관리 및 알림"
+      ],
+      "requestChecklist": "체크리스트 요청",
+      "verifiedCriteriaTitle": "Verified 인증 기준",
+      "verifiedCriteria": [
+        "문서 검증 (ANVISA/UDI/ISO)",
+        "필요 시 현지 언어 IFU",
+        "도착 시 최소 유효기간",
+        "OTIF/서비스 실적"
+      ],
+      "verifiedNote": "포털은 환자 민감 정보를 수집하지 않습니다.",
+      "rfqTitle": "견적 요청 (RFQ)",
+      "rfqSub": "공개 카탈로그가 없습니다. 요구 사항을 보내주시면 24–48시간 내 기술 비교, 검증된 문서, 제안을 드립니다.",
+      "rfqBullets": [
+        "기술 비교 (재질, 위험 등급, 호환성, 유효기간)",
+        "ANVISA • ISO 1345 • UDI • IFU/MSDS",
+        "SLA 24–48h • OTIF ≥ 95% • 리콜 및 로트 추적"
+      ],
+      "rfqHint": "제출 후 이메일로 보내드리는 보안 링크에서 파일을 첨부할 수 있습니다.",
+      "form": {
+        "name": "이름*",
+        "company": "회사*",
+        "accountType": "계정 유형",
+        "types": ["민간 병원", "공공 병원", "유통사"],
+        "email": "이메일*",
+        "phone": "전화번호*",
+        "line": "제품군*",
+        "lines": ["지혈제", "봉합사", "드릴 PPU", "Derma Pro (Beauty)"],
+        "specs": "기술 사양 / 동등 브랜드*",
+        "specsPH": "예: 위험 등급, 규격/사이즈, 흡수 특성, 호환성, 현재 등록, 참조 브랜드 등",
+        "qty": "수량",
+        "freq": "주기 (월간/분기)",
+        "deadline": "희망 납기",
+        "delivery": "납품지",
+        "reg": "규제 요구사항",
+        "regPH": "예: 등급 II/III, 필수 문서, 최소 유효기간 등",
+        "submit": "RFQ 전송",
+        "legal": "제출 시 약관에 동의하며 민감한 환자 정보를 공유하지 않음을 확인합니다.",
+        "errorRequired": "필수 항목을 완료하고 표시된 정보를 확인해주세요.",
+        "okTitle": "성공적으로 접수되었습니다",
+        "okMsg": "24–48시간 내 비교와 제안을 보내드립니다.",
+        "backTop": "맨 위로"
+      },
+      "casesTitle": "사례 및 성과",
+      "cases": [
+        {
+          "t": "학기 동안 98% 가용성",
+          "d": "사립 병원(SP) — 지혈제와 봉합사를 계획 공급으로 운영."
+        },
+        {
+          "t": "PPU 가동률 99%",
+          "d": "지역 네트워크 — 페이퍼유즈 드릴, 72시간 교체, 최소 소비 보장."
+        }
+      ],
+      "sticky": "24–48시간 견적, 기술 비교 및 검증된 문서 제공.",
+      "footer": {
+        "blurb": "컴플라이언스와 속도를 갖춘 임상 공급. 본사 🇭🇰 홍콩.",
+        "solutions": "솔루션",
+        "institutional": "회사",
+        "contact": "문의",
+        "rights": "© {year} Wonnymed. 모든 권리를 보유합니다."
+      }
+    }
+  }
+}
