+++ next.config.mjs
@@ -1,1 +1,5 @@
-export default {};
+import createNextIntlPlugin from "next-intl/plugin";
+
+const withNextIntl = createNextIntlPlugin("./i18n.js");
+
+export default withNextIntl({});
