+++ package.json
@@ -10,7 +10,8 @@
   "dependencies": {
     "next": "14.2.3",
     "react": "18.2.0",
-    "react-dom": "18.2.0"
+    "react-dom": "18.2.0",
+    "next-intl": "^3.18.0"
   },
   "devDependencies": {
     "autoprefixer": "10.4.20",
